# Conflicts:
#	pom.xml
#	src/main/java/org/example/ecommerce/common/dto/seller/CashflowDTO.java
#	src/main/java/org/example/ecommerce/controller/customer/wallet/CheckoutController.java
#	src/main/java/org/example/ecommerce/controller/seller/wallet/SellerWalletControllerAPI.java
#	src/main/java/org/example/ecommerce/entity/Customer.java
#	src/main/java/org/example/ecommerce/entity/ReviewReply.java
#	src/main/java/org/example/ecommerce/entity/Seller.java
#	src/main/java/org/example/ecommerce/repository/ReviewReplyRepository.java
#	src/main/java/org/example/ecommerce/service/ChatService.java
#	src/main/java/org/example/ecommerce/service/ChatServiceImpl.java
#	src/main/java/org/example/ecommerce/service/seller/wallet/SellerWalletService.java
#	src/main/resources/application.properties

+++ src/main/java/org/example/ecommerce/entity/Product.java
@@ -55,16 +55,16 @@ public class Product {
     private Instant createdat;
 
     @Column(name = "weight")
-    private Integer weight;
+    private Double weight;
 
     @Column(name = "length")
-    private Integer length;
+    private Double length;
 
     @Column(name = "width")
-    private Integer width;
+    private Double width;
 
     @Column(name = "height")
-    private Integer height;
+    private Double height;
 
     @Column(name = "use_variant_shipping")
     private Boolean useVariantShipping = false;
